+++ R/importAURN.R
@@ -2,16 +2,18 @@
 ##'
 ##' Function for importing hourly mean UK Automatic Urban and Rural Network
 ##' (AURN) air quality archive data files for use with the \code{openair}
-##' package. Files are imported from a remote server operated by AEA that
+##' package. Files are imported from a remote server operated by Ricardo that
 ##' provides air quality data files as R data objects.
 ##'
 ##' The \code{importAURN} function has been written to make it easy to import
-##' data from the UK AURN. AEA have provided .RData files (R workspaces) of all
-##' individual sites and years for the AURN. These files are updated on a daily
-##' basis. This approach requires a link to the Internet to work.
+##' data from the UK AURN. Ricardo have provided .RData files (R workspaces) of
+##' all individual sites and years for the AURN. These files are updated on a
+##' daily basis. This approach requires a link to the Internet to work.
 ##'
 ##' For an up to date list of available sites that can be imported, see
 ##' \code{importMeta}.
+##' 
+##' The site codes and pollutant names can be upper or lower case. 
 ##'
 ##' There are several advantages over the web portal approach where .csv files
 ##' are downloaded. First, it is quick to select a range of sites, pollutants
@@ -19,15 +21,8 @@
 ##' is very efficient as they are about four times smaller than .csv files ---
 ##' which means the data downloads quickly and saves bandwidth. Third, the
 ##' function completely avoids any need for data manipulation or setting time
-##' formats, time zones etc. Finally, it is easy to import many years of data
-##' beyond the current limit of about 64,000 lines. The final point makes it
-##' possible to download several long time series in one go. The function also
-##' has the advantage that the proper site name is imported and used in
-##' \code{openair} functions.
-##'
-##' The site codes and pollutant names can be upper or lower case. The function
-##' will issue a warning when data less than six months old is downloaded, which
-##' may not be ratified.
+##' formats, time zones etc. The function also has the advantage that the proper
+##' site name is imported and used in \code{openair} functions.
 ##'
 ##' The data are imported by stacking sites on top of one another and will have
 ##' field names \code{site}, \code{code} (the site code) and \code{pollutant}.
@@ -82,6 +77,8 @@
 ##'   data and the resulting data frames are considerably larger.
 ##' @param meta Should meta data be returned? If \code{TRUE} the site type,
 ##'   latitude and longitude are returned.
+##' @param ratified If \code{TRUE} columns are returned indicating when each
+##'   species was ratified i.e. quality-checked.
 ##' @param to_narrow By default the returned data has a column for each
 ##'   pollutant/variable. When \code{to_narrow = TRUE} the data are stacked into
 ##'   a narrow format with a column identifying the pollutant name.
@@ -115,18 +112,21 @@
 ##'
 ##' 
 importAURN <- function(site = "my1", year = 2009, pollutant = "all",
-                       hc = FALSE, meta = FALSE, to_narrow = FALSE,
-                       verbose = FALSE) {
+                       hc = FALSE, meta = FALSE, ratified = FALSE,
+                       to_narrow = FALSE, verbose = FALSE) {
 
   # For file name matching, needs to be exact
   site <- toupper(site)
+  
+  if (meta | ratified)
+    meta_data <- importMeta(source = "aurn", all = TRUE)
 
   # Create file name vector
   files <- lapply(site, function(x) paste(x, "_", year, sep = ""))
   files <- do.call(c, files)
 
-  # Donload and load data
-  thedata <- lapply(files, loadData, verbose)
+  # Download and load data
+  thedata <- lapply(files, loadData, verbose, ratified, meta_data)
   thedata <- suppressWarnings(do.call(bind_rows, thedata))
 
   # Return if no data
@@ -153,13 +153,15 @@ importAURN <- function(site = "my1", year = 2009, pollutant = "all",
     thedata <- thedata
   } else {
 
+    
     ## no hydrocarbons - therefore select conventional pollutants
     theNames <- c(
-      "date", "co", "nox", "no2", "no", "o3", "so2", "pm10", "pm2.5",
-      "v10", "v2.5", "nv10", "nv2.5", "ws", "wd", "temp", "code", "site"
+      "site", "code", "date", "co", "nox", "no2", "no", "o3", "so2", "pm10", 
+      "pm2.5", "v10", "v2.5", "nv10", "nv2.5", "ws", "wd", "temp"
     )
-
-    thedata <- thedata[, which(names(thedata) %in% theNames)]
+    
+    thedata <- select(thedata, any_of(theNames) | matches("_qc"))
+    
   }
   
   if ("temp" %in% names(thedata))
@@ -170,30 +172,31 @@ importAURN <- function(site = "my1", year = 2009, pollutant = "all",
     thedata <- thedata[, c("date", pollutant, "site", "code")]
   }
 
-
-  ## warning about recent, possibly unratified data
-  timeDiff <- difftime(Sys.time(), max(thedata$date), units = "days")
-  if (timeDiff < 180) {
-    warning("You have selected some data that is less than 6-months old.\n This most recent data is not yet ratified and may be changed\n during the QA/QC process. For complete information about the \nratification status of a data set, please use the online tool at:\n http://www.airquality.co.uk/data_and_statistics.php?action=da_1&go=Go", call. = FALSE)
-  }
-
   ## make sure it is in GMT
   attr(thedata$date, "tzone") <- "GMT"
 
   # make sure class is correct for lubridate
   class(thedata$date) <- c("POSIXct", "POSIXt")
+  
 
   if (meta) {
-    meta_data <- importMeta(source = "aurn")
+    meta_data <- distinct(meta_data, site, .keep_all = TRUE) %>% 
+      select(site, code, latitude, longitude, site_type)
     # suppress warnings about factors
     thedata <- suppressWarnings(inner_join(thedata, meta_data, by = c("code", "site")))
   }
-
+  
+ 
   if (to_narrow) {
+    
+    if (ratified) {
+      warning("Cannot re-shape if ratified is TRUE")
+      return()
+    }
     
     if (meta) {
       
-      thedata <- pivot_longer(thedata, -c(date, site, code, latitude, longitude, site.type), 
+      thedata <- pivot_longer(thedata, -c(date, site, code, latitude, longitude, site_type), 
                               names_to = "pollutant") %>% 
         arrange(site, code, pollutant, date)
       
@@ -212,7 +215,7 @@ importAURN <- function(site = "my1", year = 2009, pollutant = "all",
 
 # Define downloading and loading function
 # No export
-loadData <- function(x, verbose) {
+loadData <- function(x, verbose, ratified, meta_data) {
   tryCatch({
 
     # Download file to temp directory
@@ -240,6 +243,27 @@ loadData <- function(x, verbose) {
     
     # Reasign
     dat <- get(x)
+    
+    # add ratification information
+    if (ratified) {
+      
+      site_code <- strsplit(x, split = "_")[[1]][1]
+      
+      meta_data <- filter(meta_data, code == site_code, 
+               !variable %in% c("V10", "NV10", "V2.5", "NV2.5",
+                                "ws", "wd", "temp")) %>% 
+        select(variable, ratified_to)
+
+      for (i in 1:nrow(meta_data)) {
+
+        dat <- add_ratified(dat, variable = meta_data$variable[i],
+                            ratified_to = meta_data$ratified_to[i])
+
+      }
+
+      
+    }
+    
 
     return(dat)
     
@@ -250,4 +274,12 @@ loadData <- function(x, verbose) {
       message(x, "does not exist - ignoring that one.")
     }
   })
+}
+
+add_ratified <- function(data, variable, ratified_to) {
+
+  new_var <- paste0(variable, "_qc")
+  data <- mutate(data, {{new_var}} := ifelse(date <= ratified_to, TRUE, FALSE))
+
+  return(data)
 }