+++ R/zzz.R
@@ -7,5 +7,5 @@
                            "seslope", "slope", "weekday", "cuts", "empty",
                            "lat", "lon", "traj_len", "wkday", "date_end",
                            "validity", "process", "parameter", "date_start",
-                           "site"))
+                           "site", "where", "ratified_to", "site_type", "start_date"))
 }